qdrant setup

+++ VectorStoreConfig.go
@@ -8,7 +8,7 @@ import (
 // VectorStore - Interface for Vector Stores (already defined, ensuring it's here for context)
 type VectorStore interface {
 	Insert(vectors [][]float64, ids []string, payloads []map[string]interface{}) error
-	Search(query []float64, limit int, filters map[string]interface{}) ([]SearchResult, error)
+	Search(query []float32, limit int, filters map[string]interface{}) ([]SearchResult, error)
 	Get(vectorID string) (*SearchResult, error)
 	List(filters map[string]interface{}, limit int) ([][]SearchResult, error)
 	Update(vectorID string, vector []float64, payload map[string]interface{}) error
